+++ packages/frontend/src/app/app.module.ts
@@ -6,16 +6,31 @@ import { AppComponent } from './app.component'
 import { MapComponent } from './events/map/map.component'
 import { EventsComponent } from './events/events.component'
 import { ListComponent } from './events/list/list.component'
-import { EventService } from './core/services/event.service'
-import { HttpClientModule } from '@angular/common/http';
-import { NavComponent } from './overlay/nav/nav.component';
-import { FooterComponent } from './overlay/footer/footer.component';
-import { IconButtonComponent } from './shared/icon-button/icon-button.component'
+import { NavComponent } from './overlay/nav/nav.component'
+import { FooterComponent } from './overlay/footer/footer.component'
+import { CoreModule } from './core/core.module'
+import { SharedModule } from './shared/shared.module'
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
+import { SidenavComponent } from './overlay/sidenav/sidenav.component'
 
 @NgModule({
-  declarations: [AppComponent, MapComponent, EventsComponent, ListComponent, NavComponent, FooterComponent, IconButtonComponent],
-  imports: [BrowserModule, AppRoutingModule, HttpClientModule],
-  providers: [EventService],
+  declarations: [
+    AppComponent,
+    MapComponent,
+    EventsComponent,
+    ListComponent,
+    NavComponent,
+    FooterComponent,
+    SidenavComponent
+  ],
+  imports: [
+    BrowserModule,
+    AppRoutingModule,
+    CoreModule,
+    SharedModule,
+    BrowserAnimationsModule
+  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule {}