+++ packages/frontend/src/app/shared/material/material.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core'
+import { CommonModule } from '@angular/common'
+import { MatToolbarModule } from '@angular/material/toolbar'
+import { MatButtonModule } from '@angular/material/button'
+import { MatIconModule } from '@angular/material/icon'
+import { MatCardModule } from '@angular/material/card'
+import { MatSidenavModule } from '@angular/material/sidenav'
+import { MatListModule } from '@angular/material/list'
+import { MatDividerModule } from '@angular/material/divider'
+import { MatDialogModule } from '@angular/material/dialog'
+
+const materialModules = [
+  MatToolbarModule,
+  MatButtonModule,
+  MatIconModule,
+  MatCardModule,
+  MatSidenavModule,
+  MatListModule,
+  MatDividerModule,
+  MatDialogModule
+]
+
+@NgModule({
+  declarations: [],
+  imports: [CommonModule, ...materialModules],
+  exports: [...materialModules]
+})
+export class MaterialModule {}