+++ packages/cms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/Miwoli/jugger/compare/cms@0.4.0...cms@0.4.1) (2022-12-02)
+
+**Note:** Version bump only for package cms
+
+
+
+
+
 # [0.4.0](https://github.com/Miwoli/jugger/compare/cms@0.3.0...cms@0.4.0) (2022-11-20)